perv: change from js to tsx and restructure
