+++ package.json
@@ -6,6 +6,7 @@
     "@testing-library/jest-dom": "^4.2.4",
     "@testing-library/react": "^9.5.0",
     "@testing-library/user-event": "^7.2.1",
+    "@types/react-router-dom": "^5.1.5",
     "axios": "^0.20.0",
     "bootstrap": "^4.5.2",
     "react": "^16.13.1",
@@ -16,7 +17,8 @@
     "react-scripts": "3.4.3",
     "redux": "^4.0.5",
     "redux-devtools-extension": "^2.13.8",
-    "redux-logger": "^3.0.6"
+    "redux-logger": "^3.0.6",
+    "typescript": "^4.0.3"
   },
   "scripts": {
     "start": "react-scripts start",
