+++ src/App.tsx
@@ -1,14 +1,14 @@
 import React from "react";
 import "./App.css";
 import Navbar from "./components/Navbar";
-import Homepage from "./components/Homepage";
+import Homepage from "./pages/Homepage";
 import { BrowserRouter as Router, Route } from "react-router-dom";
 
 function App() {
   return (
     <Router>
       <div className="App">
-        <Navbar />
+         <Navbar />
         <Route path="/" exact component={Homepage} />
       </div>
     </Router>
